Create ExportAzureUserInfoCreateLocalADAccount.ps1
This script exports the information of a user in azure. Then, with this information, it creates this user's account in the on-premise domain. Since the user password is not exported, a random password is generated and assigned to the user. This password is then sent to the user as an e-mail.